更新结构
