+++ pkg/kpm/model.go
@@ -1,0 +1,13 @@
+package kpm
+
+import "github.com/orangebees/go-oneutils/GlobalStore"
+
+type CliClient struct {
+	GitStore         *GlobalStore.FileStore
+	RegistryStore    *GlobalStore.FileStore
+	WorkDir          string
+	Root             string
+	RegistryAddr     string
+	RegistryAddrPath string
+	KclVmVersion     string
+}
