+++ pkg/kpm/pkg_str.go
@@ -1,0 +1,10 @@
+package kpm
+
+type PkgString string
+
+//konfig@v0.0.1
+//konfig@v0.0.0#bbbbbb
+
+func (ps PkgString) New() {
+
+}
