+++ pkg/kpm/pkg_version.go
@@ -1,0 +1,8 @@
+package kpm
+
+// PkgVersion 包版本
+type PkgVersion string
+
+func (pv PkgVersion) GetVersion() string {
+	return ""
+}
